Fix “Your test suite must contain at least one test” problem when `it`/`test` is used without `describe`
If we use `test()` or `it()` without putting it in `describe` block,
when this reporter is active, the test will fail with confusing error:

    Your test suite must contain at least one test

Inspecting further,
I found that `allure.startCase()` will error out if `allure.startSuite` is not called before.

To fix this, I changed the notion of `suite` and `case`.
Now, 1 test file = 1 suite, regardless of `describe` blocks.
The title of each `describe` block is prepended to each case.
This matches the behavior of `jest-junit`.